feat: add docker
