--- env.example
@@ -1,7 +1,0 @@
-APP_PORT=4000
-
-### These variables are needed to make script for getting access_token works 
-AUTH_USERNAME=yourGithubLogin ### It should be your github username
-AUTH_PASSWORD=TEST_PASSWORD
-APP_URL=http://localhost:4000
-###
